+++ packages/ui/Button/Button.tsx
@@ -3,7 +3,11 @@ import * as S from './style.css';
 import createButtonVariant from './utils';
 import { iconSizes } from './constants';
 
-interface IconProps { color?: string, width: number, height: number }
+interface IconProps {
+  color?: string;
+  width: number;
+  height: number;
+}
 
 interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   children?: React.ReactNode;
@@ -11,6 +15,7 @@ interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   size?: 'sm' | 'md' | 'lg';
   theme?: 'white' | 'black' | 'blue' | 'red';
   rounded?: 'md' | 'lg';
+  variant?: 'fill' | 'outlined';
   LeftIcon?: React.ComponentType<IconProps>;
   RightIcon?: React.ComponentType<IconProps>;
 }
@@ -23,18 +28,16 @@ function Button({
   rounded = 'md',
   LeftIcon,
   RightIcon,
+  variant = 'fill',
   ...buttonElementProps
 }: ButtonProps) {
-  const variant = createButtonVariant(theme, rounded, size);
+  const style = createButtonVariant(theme, rounded, size, variant);
   const iconSize = iconSizes[size];
   return (
-    <button className={`${S.root} ${variant} ${className}`}
-      type="button"
-      {...buttonElementProps}
-    >
+    <button className={`${S.root} ${style} ${className}`} type='button' {...buttonElementProps}>
       {LeftIcon ? <LeftIcon height={iconSize} width={iconSize} /> : null}
       <span>{children}</span>
-      {(RightIcon && !LeftIcon) ? <RightIcon height={iconSize} width={iconSize} /> : null}
+      {RightIcon && !LeftIcon ? <RightIcon height={iconSize} width={iconSize} /> : null}
     </button>
   );
 }