+++ packages/ui/Button/types.ts
@@ -1,9 +1,10 @@
+export type ButtonVariant = 'fill' | 'outlined';
+
 export type ButtonColorTheme = 'white' | 'black' | 'blue' | 'red';
 
 export type ButtonBgColorStatus = 'default' | 'hover' | 'press';
 
-export type ButtonColorThemeWithStatus =
-  `${ButtonColorTheme}-${ButtonBgColorStatus}`;
+export type ButtonColorThemeWithStatus = `${ButtonVariant}-${ButtonColorTheme}-${ButtonBgColorStatus}`;
 
 export type ButtonRadiusTheme = 'md' | 'lg';
 export type ButtonSizeTheme = 'sm' | 'md' | 'lg';